+++ src/design/plone/contenttypes/restapi/services/types/get.py
@@ -108,7 +108,6 @@ FIELDSETS_ORDER = {
         "settings",
         "ownership",
         "dates",
-        "trasparenza",
     ],
     "UnitaOrganizzativa": [
         "default",
@@ -248,11 +247,17 @@ class TypesGet(BaseGet):
 
     def reorder_fieldsets(self, original):
         pt = self.request.PATH_INFO.split("/")[-1]
-        order = FIELDSETS_ORDER.get(pt, [])
+        # makea copy
+        order = [x for x in FIELDSETS_ORDER.get(pt, [])]
         if not order:
             # no match
             return original
         actual = [x["id"] for x in original]
+        portal_types = api.portal.get_tool(name="portal_types")
+        behaviors = portal_types[pt].behaviors
+        if "design.plone.contenttypes.behavior.trasparenza" in behaviors:
+            order.append("trasparenza")
+
         if set(order) != set(actual):
             # list mismatch
             raise FieldsetsMismatchError(