+++ src/design/plone/contenttypes/tests/test_ct_servizio.py
@@ -71,7 +71,6 @@ class TestServizio(unittest.TestCase):
                 "design.plone.contenttypes.behavior.argomenti",
                 "design.plone.contenttypes.behavior.additional_help_infos",
                 "collective.dexteritytextindexer",
-                "design.plone.contenttypes.behavior.trasparenza",
             ),
         )