+++ CHANGES.rst
@@ -2,7 +2,53 @@ Changelog
 =========
 
 
-1.0.3 (unreleased)
+1.0.9 (unreleased)
+------------------
+
+- Add search_sections field in control panel.
+  [cekk]
+- Can add Images into Cartella Modulistica (to be able to add image blocks in it).
+  [cekk]
+- Customizable tipologie_documento.
+  [cekk]
+
+
+1.0.8 (2021-02-19)
+------------------
+
+- Fix typo.
+  [cekk]
+
+
+1.0.7 (2021-02-19)
+------------------
+
+- Do not run dependencies when upgrading plone.app.registry.
+  [cekk]
+
+1.0.6 (2021-02-15)
+------------------
+
+- Handle Servizio tabs in both cases: with Trasparenza enabled or not.
+  [cekk]
+
+
+1.0.5 (2021-02-08)
+------------------
+
+- Disable trasparenza behavior by default.
+  [deodorhunter]
+- Remove reference limit in "persone_struttura" field.
+  [cekk]
+
+1.0.4 (2021-02-05)
+------------------
+
+- Add upgrade-step to cleanup Bando behaviors.
+  [cekk]
+
+
+1.0.3 (2021-01-20)
 ------------------
 
 - **BREAKING CHANGE** Convert RichText fields into BlocksField.